# api/src/main/java/com/mnov34/CUBES4solo/dto/EmployeeServiceDto.java
package com.mnov34.CUBES4solo.dto;

import lombok.Data;

/**
 * @author Maël NOUVEL <br>
 * 02/2025
 **/
@Data
public class EmployeeServiceDto {
    private Long id;
    private String serviceName;
}