# api/src/main/java/com/mnov34/CUBES4solo/model/ForgotPasswordRequest.java
package com.mnov34.CUBES4solo.model;

import lombok.Getter;
import lombok.Setter;

/**
 * @author Maël NOUVEL <br>
 * 02/2025
 **/
@Getter
@Setter
public class ForgotPasswordRequest {
    private String email;
}